Update listings to account for `cargo new` template update

+++ listings/ch11-writing-automated-tests/listing-11-03/src/lib.rs
@@ -1,5 +1,5 @@
 // ANCHOR: here
-pub fn add(left: usize, right: usize) -> usize {
+pub fn add(left: u64, right: u64) -> u64 {
     left + right
 }
 
@@ -12,7 +12,7 @@ mod tests {
         let result = add(2, 2);
         assert_eq!(result, 4);
     }
-    
+
     #[test]
     fn another() {
         panic!("Make this test fail");
